refactor: put back mock_client into appropriate tests
